refactor(Pub.tsx, PubDocument.tsx, useAnalytics.ts): move analytics page tracking logic to PubDocument and add collection information
