+++ client/containers/Pub/PubDocument/PubDocument.tsx
@@ -21,6 +21,9 @@ import PubInlineMenu from './PubInlineMenu';
 import PubInlineSuggestedEdits from './PubInlineSuggestedEdits';
 import PubLinkController from './PubLinkController';
 import PubMaintenanceNotice from './PubMaintenanceNotice';
+import { useAnalytics } from 'utils/analytics/useAnalytics';
+import { chooseCollectionForPub } from 'client/utils/collections';
+import { getPrimaryCollection } from 'utils/collections/primary';
 
 require('./pubDocument.scss');
 
@@ -34,7 +37,7 @@ const PubDocument = () => {
 		pubBodyState: { isReadOnly, hidePubBody },
 	} = usePubContext();
 	const { isViewingHistory } = historyData;
-	const { communityData, scopeData, featureFlags } = usePageContext();
+	const { communityData, scopeData, featureFlags, locationData } = usePageContext();
 	const pubEdgeDisplay = useFacetsQuery((F) => F.PubEdgeDisplay);
 	const { canEdit, canEditDraft } = scopeData.activePermissions;
 	const { isReviewingPub } = pubData;
@@ -46,6 +49,28 @@ const PubDocument = () => {
 	usePubHrefs({ enabled: !isReadOnly });
 
 	const showPubFileImport = (canEdit || canEditDraft) && !isReadOnly;
+
+	const uniqueCollectionIds = Array.from(
+		new Set((pubData.collectionPubs ?? []).map((cp) => cp.collectionId)),
+	);
+	const collection = chooseCollectionForPub(pubData, locationData);
+
+	const { page } = useAnalytics();
+
+	page({
+		type: 'pub',
+		communityId: pubData.communityId,
+		communityName: communityData.title,
+		title: pubData.title,
+		pubSlug: pubData.slug,
+		pubId: pubData.id,
+		pubTitle: pubData.title,
+		collectionIds: uniqueCollectionIds,
+		collectionId: collection?.id,
+		collectionTitle: collection?.title,
+		collectionSlug: collection?.slug,
+		primaryCollectionId: getPrimaryCollection(pubData?.collectionPubs)?.id,
+	});
 
 	if (hidePubBody) {
 		return null;
